Upgrade to Rails 7, importmap, Ruby 3, cssbundling
