+++ Procfile.dev
@@ -1,2 +1,2 @@
-web: bin/rails server -p 3000
+web: bin/rails server -p 3001
 css: yarn build:css --watch
