--- app/assets/stylesheets/application.sass.scss
@@ -1,1 +1,0 @@
-// Entry point for your Sass build
