+++ app/javascript/application.js
@@ -1,0 +1,3 @@
+import MicroModal from "micromodal"
+
+MicroModal.init();
