--- app/javascript/packs/stylesheets/application.scss
@@ -1,2 +1,0 @@
-@import "modal";
-@import "layout";
