--- app/javascript/packs/stylesheets/variables.scss
@@ -1,1 +1,0 @@
-$dark-red: #7a3130;
