+++ app/views/locations/_location.json.jbuilder
@@ -4,5 +4,5 @@ json.geometry do
   json.coordinates location.coords.coordinates
 end
 json.properties do
-  json.description render partial: 'locations/description.html.erb', locals: { location: location }
+  json.description render partial: 'description', formats: [:html], locals: { location: location }
 end
