+++ config/importmap.rb
@@ -1,0 +1,4 @@
+# Pin npm packages by running ./bin/importmap
+
+pin "application", preload: true
+pin "micromodal", to: "https://cdn.jsdelivr.net/gh/ghosh/Micromodal/lib/src/index.js"
